--- TUnit.Core.SourceGenerator.Tests/Bugs/1589/Tests1539.cs
@@ -1,31 +1,0 @@
-using TUnit.Core.SourceGenerator.CodeGenerators;
-
-namespace TUnit.Core.SourceGenerator.Tests.Bugs._1539;
-
-internal class Tests1589 : TestsBase<TestsGenerator>
-{
-    [Test]
-    public Task Test() => RunTest(Path.Combine(Git.RootDirectory.FullName,
-            "TUnit.TestProject",
-            "Bugs",
-            "1589",
-            "MyTests.cs"),
-        async generatedFiles =>
-        {
-            await Assert.That(generatedFiles.Length).IsEqualTo(3);
-        });
-}
-
-internal class Hooks1589 : TestsBase<TestHooksGenerator>
-{
-    [Test]
-    public Task Test() => RunTest(Path.Combine(Git.RootDirectory.FullName,
-            "TUnit.TestProject",
-            "Bugs",
-            "1589",
-            "MyTests.cs"),
-        async generatedFiles =>
-        {
-            await Assert.That(generatedFiles.Length).IsEqualTo(3);
-        });
-}
