+++ TUnit.Core.SourceGenerator.Tests/Bugs/1589/Tests1589.cs
@@ -1,0 +1,17 @@
+﻿using TUnit.Core.SourceGenerator.CodeGenerators;
+
+namespace TUnit.Core.SourceGenerator.Tests.Bugs._1589;
+
+internal class Tests1589 : TestsBase<TestsGenerator>
+{
+    [Test]
+    public Task Test() => RunTest(Path.Combine(Git.RootDirectory.FullName,
+            "TUnit.TestProject",
+            "Bugs",
+            "1589",
+            "MyTests.cs"),
+        async generatedFiles =>
+        {
+            await Assert.That(generatedFiles.Length).IsEqualTo(1);
+        });
+}
