+++ TUnit.Core.SourceGenerator.Tests/Hooks1589.Test.verified.txt
@@ -1,0 +1,62 @@
+﻿[  
+// <auto-generated/>
+#pragma warning disable
+using global::System.Linq;
+using global::System.Reflection;
+using global::System.Runtime.CompilerServices;
+using global::TUnit.Core;
+using global::TUnit.Core.Hooks;
+using global::TUnit.Core.Interfaces;
+
+namespace TUnit.SourceGenerated;
+
+[global::System.Diagnostics.StackTraceHidden]
+[global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+file partial class Hooks_BaseTests : global::TUnit.Core.Interfaces.SourceGenerator.ITestHookSource
+{
+	[global::System.Runtime.CompilerServices.ModuleInitializer]
+	public static void Initialise()
+	{
+		var instance = new Hooks_BaseTests();
+		SourceRegistrar.RegisterTestHookSource(instance);
+	}
+	public global::System.Collections.Generic.IReadOnlyList<global::TUnit.Core.Hooks.StaticHookMethod<global::TUnit.Core.TestContext>> CollectBeforeEveryTestHooks(string sessionId)
+	{
+		return
+		[
+		];
+	}
+	public global::System.Collections.Generic.IReadOnlyList<global::TUnit.Core.Hooks.StaticHookMethod<global::TUnit.Core.TestContext>> CollectAfterEveryTestHooks(string sessionId)
+	{
+		return
+		[
+		];
+	}
+	public global::System.Collections.Generic.IReadOnlyList<global::TUnit.Core.Hooks.InstanceHookMethod> CollectBeforeTestHooks(string sessionId)
+	{
+		return
+		[
+			new global::TUnit.Core.Hooks.InstanceHookMethod<global::TUnit.TestProject.Bugs._1589.BaseTests<global::TUnit.TestProject.Bugs._1589.MyFixture>>
+			{
+				MethodInfo = global::TUnit.Core.Helpers.MethodInfoRetriever.GetMethodInfo(typeof(global::TUnit.TestProject.Bugs._1589.BaseTests<global::TUnit.TestProject.Bugs._1589.MyFixture>), "Setup", 0, []),
+				Body = (classInstance, context, cancellationToken) => classInstance.Setup(),
+				HookExecutor = DefaultExecutor.Instance,
+				Order = 0,
+				MethodAttributes = [ new global::TUnit.Core.BeforeAttribute(global::TUnit.Core.HookType.Test)
+{
+    
+} ],
+				ClassAttributes = [  ],
+				AssemblyAttributes = [  ],
+			},
+		];
+	}
+	public global::System.Collections.Generic.IReadOnlyList<global::TUnit.Core.Hooks.InstanceHookMethod> CollectAfterTestHooks(string sessionId)
+	{
+		return
+		[
+		];
+	}
+}
+
+]
