+++ TUnit.Core.SourceGenerator.Tests/Tests1589.Test.verified.txt
@@ -1,0 +1,111 @@
+﻿[  
+// <auto-generated/>
+#pragma warning disable
+using global::System.Linq;
+using global::System.Reflection;
+using global::TUnit.Core;
+using global::TUnit.Core.Extensions;
+
+namespace TUnit.SourceGenerated;
+
+[global::System.Diagnostics.StackTraceHidden]
+[global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+file partial class MyTests : global::TUnit.Core.Interfaces.SourceGenerator.ITestSource
+{
+	[global::System.Runtime.CompilerServices.ModuleInitializer]
+	public static void Initialise()
+	{
+		global::TUnit.Core.SourceRegistrar.Register(new MyTests());
+	}
+	public global::System.Collections.Generic.IReadOnlyList<SourceGeneratedTestNode> CollectTests(string sessionId)
+	{
+		return Tests0(sessionId);
+	}
+	private global::System.Collections.Generic.List<SourceGeneratedTestNode> Tests0(string sessionId)
+	{
+		global::System.Collections.Generic.List<SourceGeneratedTestNode> nodes = [];
+		var classDataIndex = 0;
+		var testMethodDataIndex = 0;
+		try
+		{
+			var testClassType = typeof(global::TUnit.TestProject.Bugs._1589.MyTests);
+			var methodInfo = global::TUnit.Core.Helpers.MethodInfoRetriever.GetMethodInfo(typeof(global::TUnit.TestProject.Bugs._1589.MyTests), "Test1", 0, []);
+
+			var testBuilderContext = new global::TUnit.Core.TestBuilderContext();
+			var testBuilderContextAccessor = new global::TUnit.Core.TestBuilderContextAccessor(testBuilderContext);
+			var classArgDataGeneratorMetadata = new DataGeneratorMetadata
+{
+   Type = TUnit.Core.Enums.DataGeneratorType.Parameters,
+   TestClassType = testClassType,
+   ParameterInfos = typeof(global::TUnit.TestProject.Bugs._1589.MyTests).GetConstructors().First().GetParameters(),
+   PropertyInfo = null,
+   TestBuilderContext = testBuilderContextAccessor,
+   TestSessionId = sessionId,
+};
+			var classDataAttribute = new global::TUnit.Core.ClassDataSourceAttribute<global::TUnit.TestProject.Bugs._1589.MyFixture>()
+{
+    Shared = global::TUnit.Core.SharedType.None,
+};
+
+			var classArgGeneratedDataArray = classDataAttribute.GenerateDataSources(classArgDataGeneratorMetadata);
+
+			foreach (var classArgGeneratedDataAccessor in classArgGeneratedDataArray)
+			{
+				classDataIndex++;
+				var classArgGeneratedData = classArgGeneratedDataAccessor();
+
+				var resettableClassFactoryDelegate = () => new ResettableLazy<global::TUnit.TestProject.Bugs._1589.MyTests>(() => 
+				new global::TUnit.TestProject.Bugs._1589.MyTests(classArgGeneratedData)
+				, sessionId, testBuilderContext);
+
+				var resettableClassFactory = resettableClassFactoryDelegate();
+
+				nodes.Add(new TestMetadata<global::TUnit.TestProject.Bugs._1589.MyTests>
+				{
+					TestId = $"global::TUnit.Core.ClassDataSourceAttribute<global::TUnit.TestProject.Bugs._1589.MyFixture>:{classDataIndex}:CL-GAC0:TUnit.TestProject.Bugs._1589.MyTests(TUnit.TestProject.Bugs._1589.MyFixture).Test1:0",
+					TestClassArguments = [classArgGeneratedData],
+					TestMethodArguments = [],
+					TestClassProperties = [],
+					CurrentRepeatAttempt = 0,
+					RepeatLimit = 0,
+					MethodInfo = methodInfo,
+					ResettableClassFactory = resettableClassFactory,
+					TestMethodFactory = (classInstance, cancellationToken) => AsyncConvert.Convert(() => classInstance.Test1()),
+					TestFilePath = @"",
+					TestLineNumber = 15,
+					TestAttributes = [ new global::TUnit.Core.TestAttribute()
+{
+    
+} ],
+					ClassAttributes = [ new global::TUnit.Core.ClassDataSourceAttribute<global::TUnit.TestProject.Bugs._1589.MyFixture>()
+{
+    Shared = global::TUnit.Core.SharedType.None,
+} ],
+					AssemblyAttributes = [  ],
+					DataAttributes = [ classDataAttribute ],
+					TestBuilderContext = testBuilderContext,
+				});
+				resettableClassFactory = resettableClassFactoryDelegate();
+				testBuilderContext = new();
+				testBuilderContextAccessor.Current = testBuilderContext;
+			}
+		}
+		catch (global::System.Exception exception)
+		{
+			nodes.Add(new global::TUnit.Core.FailedInitializationTest
+			{
+				TestId = $"global::TUnit.Core.ClassDataSourceAttribute<global::TUnit.TestProject.Bugs._1589.MyFixture>:{classDataIndex}:CL-GAC0:TUnit.TestProject.Bugs._1589.MyTests(TUnit.TestProject.Bugs._1589.MyFixture).Test1:0",
+				TestClass = typeof(global::TUnit.TestProject.Bugs._1589.MyTests),
+				ReturnType = global::TUnit.Core.Helpers.MethodInfoRetriever.GetMethodInfo(typeof(global::TUnit.TestProject.Bugs._1589.MyTests), "Test1", 0, []).ReturnType,
+				ParameterTypeFullNames = [],
+				TestName = "Test1",
+				TestFilePath = @"",
+				TestLineNumber = 15,
+				Exception = exception,
+			});
+		}
+		return nodes;
+	}
+}
+
+]
